Feat: 문제 메인 사용자 정보 차트 추가
