+++ package.json
@@ -27,6 +27,7 @@
     "react-js-pagination": "^3.0.3",
     "react-router-dom": "^6.18.0",
     "react-scripts": "5.0.1",
+    "react-vis": "^1.12.1",
     "web-vitals": "^2.1.0"
   },
   "scripts": {
