+++ src/components/HeaderPage.jsx
@@ -10,7 +10,7 @@ const HeaderPage = () => {
                 <Navbar.Toggle aria-controls="basic-navbar-nav" />
                 <Navbar.Collapse id="basic-navbar-nav">
                     <Nav className="me-auto">
-                        <NavLink href="/problem">Problems</NavLink>
+                        <NavLink href="/problem/main">Problems</NavLink>
                         <NavLink href="/project/projectmain">Projects</NavLink>
                         <NavDropdown title="Community" id="basic-nav-dropdown">
                             <NavDropdown.Item href="/community/notice/NoticePage">공지사항</NavDropdown.Item>
