+++ src/components/problem/solution/SolutionPage.jsx
@@ -1,5 +1,5 @@
 import React, { useState, useCallback, useContext } from 'react';
-import { useParams, Link } from 'react-router-dom';
+import { useParams, Link, useNavigate } from 'react-router-dom';
 import axios from 'axios';
 import { useEffect } from 'react';
 
@@ -37,6 +37,8 @@ const SolutionPage = () => {
     const [complete, setComplete] = useState('0');
 
     const [executed, setExecuted] = useState(false);
+
+    const navi = useNavigate();
 
     const onChange = useCallback((val, viewUpdate) => {
         // console.log('val:', val);
@@ -137,12 +139,12 @@ const SolutionPage = () => {
     const onSubmit = async (e) => {
         e.preventDefault();
         if (executed) {
-            await axios.post("/problem/insert/solution", { problem_id, content: value, complete, user_id: sessionStorage.getItem("user_id") });
+            await axios.post("/problem/insert/solution", { problem_id, content: value, complete, language, user_id: sessionStorage.getItem("user_id") });
             setBox({
                 show: true,
                 message: "풀이가 등록되었습니다.\n풀이페이지로 이동하시겠습니까?",
                 action: () => {
-
+                    navi(`/solution/${sessionStorage.getItem("user_id")}`);
                 }
             });
         } else {
